--- a/examples/nlp/huggingface_model/text_generation/llama/quantization/weight_only/run_quant.sh
+++ b/examples/nlp/huggingface_model/text_generation/llama/quantization/weight_only/run_quant.sh
@@ -52,18 +52,18 @@ function run_tuning {
     # Check if the directory exists
     if [ ! -d "$output_model" ]; then
         # If the directory doesn't exist, create it
-	mkdir -p "$output_model"
-	echo "Created directory $output_model"
+        mkdir -p "$output_model"
+        echo "Created directory $output_model"
     fi
 
     python main.py \
             --model_path ${input_model} \
-	        --tokenizer ${tokenizer-meta-llama/Llama-2-7b-hf} \
+            --tokenizer ${tokenizer-meta-llama/Llama-2-7b-hf} \
             --output_model ${output_model} \
             --batch_size ${batch_size-1} \
             --dataset ${dataset-NeelNanda/pile-10k} \
-	        --algorithm ${algorithm-WOQ_TUNE} \
-	        --tasks ${tasks-lambada_openai} \
+            --algorithm ${algorithm-WOQ_TUNE} \
+            --tasks ${tasks-lambada_openai} \
             --tune
 }