@justinchuby 2024-05-27
All of these changes can be auto-fixed by ruff. I suggest we wait and use the tool to fix everything, once the CI pipeline is set up.